+++ alembic/versions/911ad0861ef5_squashed_baseline_wazo_23_05.py
@@ -1,0 +1,32 @@
+"""squashed baseline wazo-23.05
+
+Revision ID: 911ad0861ef5
+Revises: None
+
+"""
+
+# revision identifiers, used by Alembic.
+revision = '911ad0861ef5'
+down_revision = None
+
+import os
+
+import sqlalchemy as sa
+
+from alembic import op
+
+
+def upgrade():
+    # Read and execute the SQL dump file
+    versions_dir_path = os.path.dirname(__file__)
+    sql_file_path = os.path.join(versions_dir_path, 'baseline-2305.sql')
+
+    with open(sql_file_path) as f:
+        sql_content = f.read()
+
+    # Execute the SQL content
+    op.execute(sql_content)
+
+
+def downgrade():
+    pass
